add presure code vm2
