+++ cmd/tx-factory/src/lib.rs
@@ -1,2 +1,4 @@
+pub mod pressure_vm1;
+pub mod pressure_vm2;
 pub mod txn_generator;
 pub mod vm2_txn_lib;
