+++ kube/manifest/starcoin-halley.yaml
@@ -11,7 +11,7 @@ spec:
     matchLabels:
       app: starcoin
   serviceName: starcoin-svc
-  replicas: 1
+  replicas: 3
   template:
     metadata:
       name: starcoin
@@ -23,27 +23,14 @@ spec:
         starcoin/node-pool: seed-pool
       containers:
       - name: starcoin
-        image: starcoin/starcoin:halley24_force_upgrade_v11
+        image: starcoin/starcoin:sha-6a5e4e8
         imagePullPolicy: Always
         command:
           - bash
           - -c
         args:
-          -
-            rm -rf /sc-data/halley/starcoin.ipc /sc-data/halley/starcoindb/db/starcoindb/LOCK;
-            id=$(echo -e $POD_NAME|awk -F'-' '{print $2}') && IFS='; ' read -r -a node_keys <<< $NODE_KEYS &&
-            node_key=${node_keys[$id]};
-            if [ ! -z $node_key ]; then
-              node_key_flag="--node-key ${node_key}";
-            fi;
-            /starcoin/starcoin -n halley -d /sc-data --discover-local=true $node_key_flag;
-            ret=$?;
-            if [ $ret -ne 0 ]; then
-              echo "Node start fail, try to remove config and data.";
-              rm /sc-data/halley/config.toml;
-              rm /sc-data/halley/genesis_config.json;
-              rm -rf /sc-data/halley/ &>/dev/null;
-            fi;
+          - |
+            rm -rf /sc-data/halley; rm -rf /sc-data/halley/starcoin.ipc /sc-data/halley/starcoindb/db/starcoindb/LOCK /sc-data/halley/genesis_config.json; id=$(echo -e $POD_NAME | awk -F'-' '{print $2}'); IFS='; ' read -r -a node_keys <<< "$NODE_KEYS"; node_key=${node_keys[$id]}; node_key_flag=""; if [ -n "$node_key" ]; then node_key_flag="--node-key ${node_key}"; fi; miner_thread_flag=""; { case "$POD_NAME" in starcoin-0) miner_thread_flag="--miner-thread=1" ;; starcoin-1) miner_thread_flag="--miner-thread=2" ;; starcoin-2) miner_thread_flag="--miner-thread=3" ;; *) miner_thread_flag="" ;; esac; }; /starcoin/starcoin -n halley -d /sc-data --p2prpc-default-global-api-quota 9000/s --p2prpc-custom-user-api-quota get_header_by_hash=9000/s --p2prpc-custom-user-api-quota get_headers_by_hash=9000/s --p2prpc-custom-user-api-quota info=9000/s --p2prpc-custom-user-api-quota get_block_by_hash=9000/s --p2prpc-custom-user-api-quota get_block_ids=9000/s --p2prpc-custom-user-api-quota get_blocks=9000/s --jsonrpc-default-global-api-quota 9000/s --jsonrpc-custom-user-api-quota chain.get_headers_by_hash=9000/s --jsonrpc-custom-user-api-quota chain.get_header_by_hash=9000/s --jsonrpc-custom-user-api-quota chain.info=9000/s --jsonrpc-custom-user-api-quota chain.get_block_by_hash=9000/s --jsonrpc-custom-user-api-quota chain.get_block_ids=9000/s --jsonrpc-custom-user-api-quota chain.get_blocks=9000/s --min-peers-to-propagate 512 --max-peers-to-propagate 1024 --max-outgoing-peers 512 --max-incoming-peers 512 --txpool-tx-propagate-interval 1 --discover-local=true $node_key_flag $miner_thread_flag
         ports:
           - containerPort: 9840
             hostPort: 9840
@@ -63,18 +50,6 @@ spec:
         livenessProbe:
           tcpSocket:
             port: 9840
-          initialDelaySeconds: 10
-          periodSeconds: 5
-          timeoutSeconds: 2
-          failureThreshold: 3
-          successThreshold: 1
-        readinessProbe:
-          exec:
-            command:
-             - sh
-             - -c
-             - >-
-                 /starcoin/starcoin -n halley -d /sc-data node sync status|grep Synchronized
           initialDelaySeconds: 10
           periodSeconds: 5
           timeoutSeconds: 2
